+++ src/main/kotlin/no/nav/tms/utbetalingsoversikt/api/ytelse/domain/transformer/HovedytelseTransformer.kt
@@ -1,6 +1,7 @@
 package no.nav.tms.utbetalingsoversikt.api.ytelse.domain.transformer
 
 import io.ktor.util.*
+import no.nav.tms.utbetalingsoversikt.api.utbetaling.YtelseIdUtil
 import no.nav.tms.utbetalingsoversikt.api.ytelse.domain.external.AktoerEkstern
 import no.nav.tms.utbetalingsoversikt.api.ytelse.domain.external.PeriodeEkstern
 import no.nav.tms.utbetalingsoversikt.api.ytelse.domain.external.UtbetalingEkstern
@@ -17,7 +18,7 @@ object HovedytelseTransformer {
     fun toHovedYtelse(utbetaling: UtbetalingEkstern): List<Hovedytelse> {
         return utbetaling.ytelseListe.map { ytelseEkstern ->
             Hovedytelse(
-                id = createImplicitId(utbetaling, ytelseEkstern),
+                id = YtelseIdUtil.calculateId(utbetaling, ytelseEkstern),
                 rettighetshaver = createRettighetshaver(ytelseEkstern.rettighetshaver),
                 ytelse = ytelseEkstern.ytelsestype?: "",
                 status = utbetaling.utbetalingsstatus,
@@ -32,18 +33,6 @@ object HovedytelseTransformer {
                 melding = utbetaling.utbetalingsmelding?: ""
             )
         }
-    }
-
-    private fun createImplicitId(utbetaling: UtbetalingEkstern, ytelse: YtelseEkstern): String {
-        val datePart = LocalDate.parse(utbetaling.posteringsdato)
-            .toEpochDay()
-
-        val contentPart = ytelse.hashCode()
-            .toString()
-
-        return "$datePart-$contentPart"
-            .toByteArray()
-            .encodeBase64()
     }
 
     private fun createRettighetshaver(aktoer: AktoerEkstern): Rettighetshaver {
